Adjust printf formats for 64bit time_t on 32bit systems
This fixes a crash when time_t is 64bit and also
fL::wqixes format specifier mismatch warnings as well while here

e.g.
warning: format '%ld' expects argument of type 'long int', but argument 2 has type 'time_t'

+++ src/misc.c
@@ -1512,11 +1512,11 @@ cr_append_pid_and_datetime(const char *str, const char *suffix)
     gettimeofday(&tv, NULL);
     timeinfo = localtime (&(tv.tv_sec));
     strftime(datetime, 80, "%Y%m%d%H%M%S", timeinfo);
-    gchar *result = g_strdup_printf("%s%jd.%s.%ld%s",
+    gchar *result = g_strdup_printf("%s%jd.%s.%jd%s",
                                     str ? str : "",
                                     (intmax_t) getpid(),
                                     datetime,
-                                    tv.tv_usec,
+                                    (intmax_t) tv.tv_usec,
                                     suffix ? suffix : "");
     return result;
 }
